refactor(types): move User and CurrentUser interfaces to @sanity/types
