+++ packages/@sanity/base/src/datastores/user/hooks.ts
@@ -4,8 +4,8 @@
 import userStore from 'part:@sanity/base/user'
 
 import {useMemo} from 'react'
+import {CurrentUser, User} from '@sanity/types'
 import {LoadableState, useLoadable} from '../../util/useLoadable'
-import {CurrentUser, User} from './types'
 
 export function useUser(userId: string): LoadableState<User | undefined> {
   return useLoadable(useMemo(() => userStore.observable.getUser(userId), [userId]))
