+++ packages/@sanity/base/src/presence/types.ts
@@ -1,5 +1,5 @@
-import {Path} from '@sanity/types'
-import {Session, Status, User} from '../datastores/presence/types'
+import {Path, User} from '@sanity/types'
+import {Session, Status} from '../datastores/presence/types'
 
 export type Position = 'top' | 'bottom' | 'inside' | null
 export type Size = 'xsmall' | 'small' | 'medium'
