+++ packages/@sanity/types/src/index.ts
@@ -9,3 +9,4 @@ export * from './validation'
 export * from './reference'
 export * from './mutations'
 export * from './images'
+export * from './user'
